requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# Data science and ML
numpy>=1.20.0
pyyaml>=5.4.0
requests>=2.25.0
